added linkedin logo
