+++ modules/basic-landing-page-components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@idahoedokpayi/basic-landing-page-components",
-  "version": "0.1.5",
+  "version": "0.1.6",
   "description": "React Components module for basic-landing-page-components",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
