+++ modules/basic-landing-page-images/index.js
@@ -1,7 +1,8 @@
+import linkedin from "./images/LI-In-Bug.png";
 import boats from "./images/boats.jpg";
 import bg from "./images/business-district.jpg";
 import pic1 from "./images/pic1.jpg";
 import pic2 from "./images/pic2.jpg";
 import logo from "./images/trident-color.svg";
 
-export { bg, boats, logo, pic1, pic2 };
+export { bg, boats, linkedin, logo, pic1, pic2 };
