+++ themes/bootstrap-album/content/album/bootstrapcover.mdx
@@ -1,0 +1,6 @@
+---
+title: "Gatsby Bootstrap Cover Theme"
+date: "2020-05-15"
+featuredImage: "../images/bootstrapcover.jpg"
+---
+A screen shot of the Gatsby Product Cover Theme built in Typescript. Based on [Bootstrap Cover Example](https://getbootstrap.com/docs/4.5/examples/cover/).
