+++ themes/bootstrap-album/content/album/bootstrapproduct.mdx
@@ -1,0 +1,6 @@
+---
+title: "Gatsby Bootstrap Product Landing Page Theme"
+date: "2020-05-15"
+featuredImage: "../images/bootstrapproduct.jpg"
+---
+A screen shot of the Gatsby Product Landing Page Theme built in Typescript. Based on [Bootstrap Product Example](https://getbootstrap.com/docs/4.5/examples/product/).
