+++ themes/bootstrap-album/package.json
@@ -47,7 +47,7 @@
   },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
-    "@idahoedokpayi/basic-landing-page-components": "^0.1.5",
+    "@idahoedokpayi/basic-landing-page-components": "^0.1.6",
     "@idahoedokpayi/basic-landing-page-svg": "0.0.1",
     "@mdx-js/mdx": "^1.6.1",
     "@mdx-js/react": "^1.6.1",
